Upgrade base (#32)
* Properly use Go modules

For whatever reason, the Go 1.10-like support via a separate script to pre-retrieve and cache packages is present. This upgrades from Go 1.11 to Go 1.15 and properly utilizes modules.

The get.sh additionally directly contradicts the README. What was I thinking? We may never know..

* Put example in proper location

The `./config` directory is bound as a volume within the compose definitions file. It's unknown why this file was moved. Mail-Go relies on the config dir existing to create and retain its salt.

* Fix mlid validation issues